Exception method for Not Existing ID (for Delete and Put Mapping)

+++ DatabaseManagementSystem/src/main/java/database/management/system/databasemanagementsystem/service/impl/ServiceImpl.java
@@ -1,5 +1,6 @@
 package database.management.system.databasemanagementsystem.service.impl;
 
+import database.management.system.databasemanagementsystem.exception.NotExistingID;
 import database.management.system.databasemanagementsystem.exception.RepeatedIdErrorResponse;
 import database.management.system.databasemanagementsystem.model.Doctor;
 import database.management.system.databasemanagementsystem.repository.DoctorRepository;
@@ -36,25 +37,33 @@ public class ServiceImpl implements ServiceRep {
     }
 
     @Override
-    public Doctor deleteDoctorByID(Integer id) {
+    public Doctor deleteDoctorByID(Integer id) throws NotExistingID {
         Doctor doctorToBeRemoved = doctorRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Employee not exist with id: " + id));
-        doctorRepository.deleteById(id);
-        return doctorToBeRemoved;
+        if (doctorRepository.existsById(doctorToBeRemoved.getId())) {
+            doctorRepository.deleteById(id);
+            return doctorToBeRemoved;
+        } else {
+            throw new NotExistingID();
+        }
     }
 
     @Override
-    public Doctor updateDoctorByID(Integer id, Doctor updatedDoctor) {
-        return doctorRepository.findById(id)
-            .map(doctor -> {
-                doctor.setFullName(updatedDoctor.getFullName());
-                doctor.setDateOfBirth(updatedDoctor.getDateOfBirth());
-                doctor.setNation(updatedDoctor.getNation());
-                doctor.setHospital(updatedDoctor.getHospital());
-                doctor.setSpecialization(updatedDoctor.getSpecialization());
-                doctor.setFreeDay(updatedDoctor.getFreeDay());
-                return doctorRepository.save(doctor);
-            })
-            .orElseThrow(() -> new RuntimeException("Doctor not found with id " + id));
+    public Doctor updateDoctorByID(Integer id, Doctor updatedDoctor) throws NotExistingID {
+        if (doctorRepository.existsById(updatedDoctor.getId())) {
+            return doctorRepository.findById(id)
+                    .map(doctor -> {
+                        doctor.setFullName(updatedDoctor.getFullName());
+                        doctor.setDateOfBirth(updatedDoctor.getDateOfBirth());
+                        doctor.setNation(updatedDoctor.getNation());
+                        doctor.setHospital(updatedDoctor.getHospital());
+                        doctor.setSpecialization(updatedDoctor.getSpecialization());
+                        doctor.setFreeDay(updatedDoctor.getFreeDay());
+                        return doctorRepository.save(doctor);
+                    })
+                    .orElseThrow(() -> new RuntimeException("Doctor not found with id " + id));
+        } else {
+            throw new NotExistingID();
+        }
     }
 }
